fix(rerendering): reduce the re-rendering of the viewport grid
