+++ extensions/cornerstone/src/components/CinePlayer/index.ts
@@ -1,0 +1,3 @@
+import CinePlayer from './CinePlayer';
+
+export default CinePlayer;
